+++ R/utils_draws.R
@@ -3,17 +3,21 @@ brm_sep <- function() {
 }
 
 gsub_group <- function(names) {
-  out <- strsplit(names, split = "|", fixed = TRUE)
+  out <- strsplit(names, split = brm_sep(), fixed = TRUE)
   as.character(lapply(out, function(x) x[[1L]]))
 }
 
 gsub_time <- function(names) {
-  out <- strsplit(names, split = "|", fixed = TRUE)
+  out <- strsplit(names, split = brm_sep(), fixed = TRUE)
   as.character(lapply(out, function(x) x[[2L]]))
 }
 
 name_marginal <- function(group, time) {
   sprintf("%s%s%s", group, brm_sep(), time)
+}
+
+name_marginal_subgroup <- function(group, subgroup, time) {
+  sprintf("%s%s%s%s%s", group, brm_sep(), subgroup, brm_sep(), time)
 }
 
 names_group <- function(draws) {